+++ src/components/common/PermissionBlock/PermissionBlock.module.css
@@ -102,3 +102,7 @@
     }
   }
 }
+
+.loginTitle {
+  margin-bottom: 34px;
+}
